requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
awsebcli==3.20.3
azure-cognitiveservices-speech==1.21.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
azure-core==1.23.1
azure-storage-blob==12.11.0
bcrypt==3.2.0
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.1
django-autocomplete-light==3.9.1
django-cors-headers==3.11.0
django-heroku==0.3.1
django-import-export==2.7.1
django-storages==1.12.3
django-storages-azure==1.6.8
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
fastapi==0.73.0
future==0.16.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httpie==3.0.2
httptools==0.3.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
isodate==0.6.1
jmespath==0.10.0
jsonify==0.5
jsonschema==3.2.0
MarkupPy==1.14
msrest==0.6.21
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.2
paramiko==2.9.2
pathspec==0.9.0
Pillow==9.0.0
priority==2.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyNaCl==1.5.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
random-password-generator==2.2.0
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
stripe==3.3.0
tablib==3.2.0
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
uvloop==0.16.0
watchgod==0.7
wcwidth==0.1.9
websocket-client==0.59.0
websockets==10.1
whitenoise==6.0.0
wsproto==1.0.0
xlrd==2.0.1
xlwt==1.3.0
